LPS-105254 Update liferay-npm-scripts to v15.0.0
Prior to this, we formatted JS inside JSP using Prettier, but that only
enforces presentational aspects.

Now, we run JS inside JSP through ESLint too, and that enables us to
catch a lot of issues, particularly ones that affect IE11. See
LPS-105255 and the related PR where I fixed all of the issues detected
by this new version of liferay-npm-scripts:

https://github.com/brianchandotcom/liferay-portal/pull/81568