LPS-104983 Move network information to NetworkContext provider
Network information is now in its own context to minimize infinite looping in the Store. One downside to this is that everything here still seems manual, whenever you have to create a thunk you will need to pass the fetcher to the service, which may seem strange but if it wasn't done we would still have to keep doing it to pass the callback of `onServiceStatus` for each thunk.

Another option for this is to go with the idea of ​​the `useService` hook that will dock with `useThunk`, so any thunk will have `fetcher` visibility, making it less painful to call the network context every time you use a thunk .